#4 Setup CI using github actions

+++ runtest.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
 
+set -eE
+
 function cleanup() {
+  if [ -n "$SERVER_PID" ]; then kill $SERVER_PID; fi
   rm -Rf assets
 }
 
 trap cleanup EXIT
+trap cleanup ERR
 
 yarn webpack
 python -m http.server 8000 &
 SERVER_PID=$!
 
-pytest --ignore=node_modules -s
-
-kill $SERVER_PID
+xvfb-run -a pytest --ignore=node_modules -s
